.github/skills/pyrit-prompt-validation/requirements.txt
# Real PyRIT Framework - Latest version (0.10.0+)
# Security testing for AI system prompts

# Core PyRIT - latest version
pyrit>=0.10.0

# Azure identity for Azure OpenAI
azure-identity>=1.15.0

# Environment management
python-dotenv>=1.0.0

# Console output formatting
colorama>=0.4.6

# Optional for advanced features
pyyaml>=6.0.1
jinja2>=3.1.3

# Development and type checking
mypy>=1.8.0  # Static type checker
pytest>=8.0.0  # Testing framework
pytest-asyncio>=0.23.0  # Async test support
